Need to work on deleting reservation

+++ src/main/java/com/hairshop_reservation_system/springboot_hairshop_reservation_system/service/impl/ReservationServiceImpl.java
@@ -50,7 +50,7 @@ public class ReservationServiceImpl implements ReservationService {
     @Override
     public String deleteReservationById(String reservationId) {
         try{
-            reservationRepository.deleteReservationById(Long.parseLong(reservationId));
+            reservationRepository.deleteReservationById(Integer.parseInt(reservationId));
         }catch (Exception e){
             e.printStackTrace();
             return "failed";
